feat(assets/, components/): Created assets and components folder for navbar The assets folder contains a tailwind.js which is used in the app.py to use tailwind cdn to allow us to use tailwindcss without having to install tailwind packages locally.
The components folder holds the sidebar.py which is the navbar itself. It is dynamic in terms of the tabs. The names and the pages/url can be dynamically inputted to update the navbar.

refactor(app.py): Created navbar for the webpage.
Use sidebar.py in the app.py to add nav bar. Created callbacks to handle toggle of the sidebar and the colors of the tabs in the sidebar.